move canAccess into @wdio/utils
