+++ packages/wdio-utils/src/utils.js
@@ -1,3 +1,5 @@
+import fs from 'fs'
+
 const SCREENSHOT_REPLACEMENT = '"<Screenshot[base64]>"'
 
 /**
@@ -220,4 +222,22 @@ export function isBase64(str) {
     return firstPaddingChar === -1 ||
       firstPaddingChar === len - 1 ||
       (firstPaddingChar === len - 2 && str[len - 1] === '=')
+}
+
+/**
+ * Helper utility to check file access
+ * @param {String} file file to check access for
+ * @return              true if file can be accessed
+ */
+export const canAccess = (file) => {
+    if (!file) {
+        return false
+    }
+
+    try {
+        fs.accessSync(file)
+        return true
+    } catch (e) {
+        return false
+    }
 }
